Use double for resamplers
Former-commit-id: 6c024e150c3521c9a8f6ac12da96abfd7b4d1be7
Former-commit-id: 598e6e70606a95d16c6205d4eb04a960842246aa
Former-commit-id: 2c3e00cf2db86aaf7eedcbca3b7eef382d689885

+++ src/ImageProcessorCore/Samplers/Resamplers/BoxResampler.cs
@@ -12,10 +12,10 @@ namespace ImageProcessorCore
     public class BoxResampler : IResampler
     {
         /// <inheritdoc/>
-        public float Radius => 0.5F;
+        public double Radius => 0.5d;
 
         /// <inheritdoc/>
-        public float GetValue(float x)
+        public double GetValue(double x)
         {
             if (x > -0.5 && x <= 0.5)
             {
